refactor(cli): move URL JSON serialization to its owner
