+++ src/run/flows/url/summary.ts
@@ -17,7 +17,6 @@ import {
   resolveUrlSummaryExecution,
   type UrlSummaryResolution,
 } from "../../../engine/web-summary.js";
-import { buildRunJsonEnv } from "../../../shared/run-api-status.js";
 import { buildExtractFinishLabel, writeFinishLine } from "../../finish-line.js";
 import { writeVerbose } from "../../logging.js";
 import { prepareMarkdownForTerminal } from "../../markdown.js";
@@ -26,7 +25,7 @@ import type { UrlExtractionUi } from "./extract.js";
 import type { SlidesTerminalOutput } from "./slides-output.js";
 import { formatSourceMetricsHeader } from "./source-metrics.js";
 import { buildFinishExtras, pickModelForFinishLine } from "./summary-finish.js";
-import { buildUrlJsonInput } from "./summary-json.js";
+import { writeUrlJsonOutput } from "./summary-json.js";
 import type { UrlFlowContext } from "./types.js";
 
 type SlidesResult = Awaited<
@@ -43,55 +42,6 @@ type ExtractedUrlOutputArgs = {
   transcriptionCostLabel: string | null;
   slides?: SlidesResult | null;
 };
-
-async function writeUrlJsonOutput({
-  ctx,
-  url,
-  extracted,
-  effectiveMarkdownMode,
-  prompt,
-  slides,
-  summary,
-  llm,
-}: {
-  ctx: UrlFlowContext;
-  url: string;
-  extracted: ExtractedLinkContent;
-  effectiveMarkdownMode: "off" | "auto" | "llm" | "readability";
-  prompt: string;
-  slides?: SlidesResult | null;
-  summary: string | null;
-  llm: {
-    provider: string;
-    model: string;
-    maxCompletionTokens: number | null;
-    strategy: "single";
-  } | null;
-}): Promise<RunMetricsReport | null> {
-  const { io, flags, model, hooks } = ctx;
-  hooks.clearProgressForStdout();
-  const finishReport = flags.shouldComputeReport ? await hooks.buildReport() : null;
-  const payload = {
-    input: {
-      ...buildUrlJsonInput({
-        flags,
-        url,
-        effectiveMarkdownMode,
-        modelLabel: model.requestedModelLabel,
-      }),
-    },
-    env: buildRunJsonEnv(model.apiStatus),
-    extracted,
-    slides,
-    prompt,
-    llm,
-    metrics: flags.metricsEnabled ? finishReport : null,
-    summary,
-  };
-  io.stdout.write(`${JSON.stringify(payload, null, 2)}\n`);
-  hooks.restoreProgressAfterStdout?.();
-  return finishReport;
-}
 
 async function writeUrlMetricsFinishLine({
   ctx,
